Add unit test for file read and write
